Remove unnecessary additions to build system. Provide a production build process.
