+++ index.ejs
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+  <meta charset="UTF-8" />
+  <title>Voyager 2</title>
+</head>
+
+<body>
+  <div id="root"></div>
+</body>
+
+</html>
