+++ index.html
@@ -4,11 +4,20 @@
 <head>
   <meta charset="UTF-8" />
   <title>Voyager 2</title>
+  <link href="./dist/style.css" rel="stylesheet">
 
 </head>
 
 <body>
   <div id="root"></div>
+
+  <!-- Dependencies -->
+  <script src="./node_modules/react/dist/react.js"></script>
+  <script src="./node_modules/react-dom/dist/react-dom.js"></script>
+
+  <!-- Main -->
+  <script src="./dist/vendor.js"></script>
+  <script src="./dist/bundle.js"></script>
 
 </body>
