+++ package.json
@@ -17,7 +17,7 @@
   },
   "license": "BSD-3-Clause",
   "scripts": {
-    "start": "node scripts/start.js",
+    "start": "webpack-dev-server --open --config config/webpack.config.dev.js",
     "build": "node scripts/build.js",
     "clean": "find -E dist  -regex '.*\\.(js|js.map|d.ts)' -delete",
     "lint": "tslint -c tslint.json 'src/**/*.ts' 'src/**/*.tsx'",
@@ -54,7 +54,6 @@
     "vega-lite": "2.0.0-alpha.7"
   },
   "devDependencies": {
-    "@timer/detect-port": "^1.1.3",
     "@types/classnames": "^0.0.32",
     "@types/jest": "^18.1.1",
     "@types/react": "^15.0.8",
@@ -67,21 +66,17 @@
     "@types/redux-logger": "^3.0.0",
     "address": "^1.0.1",
     "autoprefixer": "^6.7.6",
-    "babel-loader": "^7.0.0",
     "case-sensitive-paths-webpack-plugin": "^2.0.0",
     "codecov": "^1.0.1",
     "css-loader": "^0.26.1",
-    "eslint-loader": "^1.7.1",
     "extract-text-webpack-plugin": "beta",
     "file-loader": "^0.10.1",
     "fs-extra": "^3.0.1",
     "html-webpack-plugin": "^2.28.0",
-    "inquirer": "^3.0.6",
     "jest": "~18",
     "node-sass": "^4.5.0",
     "postcss-loader": "^1.3.3",
-    "react-dev-utils": "^1.0.0-alpha.e280254",
-    "react-error-overlay": "^1.0.0-alpha.e280254",
+    "react-dev-utils": "^0.5.2",
     "sass-loader": "^5.0.1",
     "source-map-loader": "^0.1.6",
     "style-loader": "^0.13.1",
