+++ src/constants.ts
@@ -17,6 +17,8 @@ export enum FieldParentType {
   FIELD_LIST
 };
 
+const BASE_DATA_DIR = (process.env.NODE_ENV === 'production') ? 'datasets/' : 'node_modules/vega-datasets/';
+
 export const DEFAULT_DATASETS = [{
   name: '-',
   url: undefined
@@ -80,6 +82,6 @@ export const DEFAULT_DATASETS = [{
 }].map(dataset => {
   return {
     ...dataset,
-    url: 'node_modules/vega-datasets/' + dataset.url
+    url: BASE_DATA_DIR + dataset.url
   };
 });
